Rename 'rquest-omop-agent' to 'rquest-omop-worker'
