--- app/rquest-omop-agent/README.md
@@ -1,1 +1,0 @@
-# RQuest OMOP agent
